refactor: components path & directories, types, names, styles
