+++ src/entities/Camper/index.ts
@@ -1,6 +1,8 @@
-import { type ICamper, CamperRole, type CamperSocial } from './model/types/camper.types.ts';
+import { type ICamper, CamperRole, type CamperSocial } from './model/types/camper.types';
+import useCampers from './model/services/useCampers/useCampers';
 
 export {
+  useCampers,
   ICamper,
   CamperRole,
   CamperSocial,
