+++ src/entities/User/index.ts
@@ -1,8 +1,8 @@
 import UserAvatar from './ui/UserAvatar/UserAvatar';
 import UserSignInForm from './ui/UserSignInForm/UserSignInForm';
 import { type SignInFormData } from './ui/UserSignInForm/UserSignInForm.types';
-import OwnerSignUpForm from './ui/OwnerSignUpForm/OwnerSignUpForm';
-import { type IInputsData } from './ui/OwnerSignUpForm/OwnerSignUpForm.types';
+import TCOSignUpForm from './ui/TCOSignUpForm/TCOSignUpForm';
+import { type IInputsData } from './ui/TCOSignUpForm/TCOSignUpForm.types';
 import useAuth, { type IIDToken } from './model/services/useAuth/useAuth';
 import { type IUserAvatar } from './ui/UserAvatar/UserAvatar';
 
@@ -10,7 +10,7 @@ export {
   UserAvatar,
   UserSignInForm,
   SignInFormData,
-  OwnerSignUpForm,
+  TCOSignUpForm,
   IInputsData,
   useAuth,
   IUserAvatar,
