+++ src/entities/User/ui/UserSignInForm/UserSignInForm.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 import { Link } from 'react-router-dom';
 import { Form, Formik } from 'formik';
 
-import { CustomInput } from 'shared/ui/CustomInput/CustomInput';
+import CustomInput from 'shared/ui/CustomInput/CustomInput';
 import Button from 'shared/ui/Button/Button';
 import Icon from 'shared/ui/Icon/Icon';
