+++ src/features/InviteUserModal/index.ts
@@ -1,0 +1,3 @@
+import InviteUserModal from './ui/InviteUserModal';
+
+export default InviteUserModal;
