+++ src/features/UserDetailsModal/index.ts
@@ -1,0 +1,3 @@
+import UserDetailsModal from './ui/UserDetailsModal';
+
+export default UserDetailsModal;
