+++ src/features/UserDetailsModal/ui/UserDetailsModal.module.scss
@@ -1,4 +1,4 @@
-@import 'breakpoints';
+@import '../../../app/styles/breakpoints';
 
 .details {
   display: grid;
