--- src/pages/FunnelPage/ui/InviteMember/InviteMember.module.scss
@@ -1,55 +1,0 @@
-@import 'breakpoints';
-
-.input {
-  border-radius: 10px;
-  border: 2px solid var(--color-input-orange);
-  background: rgba(255, 255, 255, 0.02);
-  backdrop-filter: blur(4px);
-  padding: 10px 20px;
-  font-size: 14px;
-  width: 100%;
-
-  &::placeholder {
-    color: var(--color-neutral);
-  }
-}
-.radioGroup {
-  display: flex;
-  gap: 15px;
-  &:hover {
-    cursor: pointer;
-  }
-  & label {
-    font-size: var(--font-size-s);
-
-    display: flex;
-    align-items: center;
-    gap: 7px;
-  }
-}
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 23px;
-  & button {
-    margin: 0 auto;
-  }
-}
-.wrapper {
-  width: 490px;
-  max-width: 100%;
-  & h2 {
-    font-size: var(--font-size-m);
-    margin-bottom: 10px;
-    font-weight: 500;
-  }
-  & p {
-    width: 80%;
-    font-size: var(--font-size-s);
-    margin-bottom: 20px;
-
-    @media (width < $md) {
-      width: 100%;
-    }
-  }
-}
