+++ src/pages/SignUpCamperPage/index.ts
@@ -1,3 +1,3 @@
-import { SignUpCamperPage } from './ui/SignUpCamperPage/SignUpCamperPage.async';
+import { SignUpCamperPage } from './ui/SignUpCamperPage.async';
 
 export default SignUpCamperPage;
