+++ src/pages/SignUpCamperPage/ui/SignUpCamperPage.tsx
