--- src/pages/SignUpOwnerPage/index.ts
@@ -1,3 +1,0 @@
-import { SignUpOwnerPage } from './ui/SignUpOwnerPage/SignUpOwnerPage.async';
-
-export default SignUpOwnerPage;
