--- src/pages/SignUpOwnerPage/ui/SignUpOwnerPage/SignUpOwnerPage.async.tsx
@@ -1,5 +1,0 @@
-import { lazy } from 'react';
-
-const SignUpOwnerPageAsync = lazy(async () => import('./SignUpOwnerPage'));
-
-export { SignUpOwnerPageAsync as SignUpOwnerPage };
