+++ src/pages/SignUpTCOPage/index.ts
@@ -1,0 +1,3 @@
+import { SignUpTCOPage } from './ui/SignUpTCOPage.async';
+
+export default SignUpTCOPage;
