+++ src/pages/SignUpTCOPage/ui/SignUpTCOPage.async.tsx
@@ -1,0 +1,5 @@
+import { lazy } from 'react';
+
+const SignUpTCOPageAsync = lazy(async () => import('./SignUpTCOPage'));
+
+export { SignUpTCOPageAsync as SignUpTCOPage };
