+++ src/shared/ui/CustomInput/CustomInput.module.scss
@@ -4,7 +4,6 @@
   background: rgba(255, 255, 255, 0.02);
   backdrop-filter: blur(4px);
   padding: 10px 20px;
-  font-family: Poppins;
   font-size: 14px;
   width: 100%;
 
@@ -16,7 +15,7 @@
   display: flex;
   align-items: center;
   gap: 5px;
-  font-size: var(--font-size-xs);
+  font: var(--font-xs);
 }
 .valid {
   color: #038f00;
@@ -31,7 +30,7 @@
   flex-direction: column;
   gap: 5px;
   & p {
-    font-size: 14px;
+    font: var(--font-s);
     font-weight: 500;
   }
 }
@@ -42,10 +41,10 @@
   transform: translateY(-50%);
   cursor: pointer;
   z-index: 100;
-  background-color: none;
+  background-color: transparent;
   border: none;
 }
 .tooltip {
   z-index: 100;
-  box-shadow: 0px 4px 20px 0px rgba(0, 0, 0, 0.25);
+  box-shadow: 0 4px 20px 0 rgba(0, 0, 0, 0.25);
 }
