+++ src/shared/ui/CustomRadio/CustomRadio.module.scss
@@ -1,0 +1,49 @@
+.radio {
+	position: relative;
+	cursor: pointer;
+
+	@media (hover: hover) {
+		&:hover {
+			.label:before {
+				background-color: var(--color-ruby-dark);
+				box-shadow: inset 0 0 0 4px var(--color-secondary-bg);
+			}
+		}
+	}
+
+	.input {
+		position: absolute;
+		top: 0;
+		left: 0;
+		width: 0;
+		height: 0;
+		opacity: 0;
+		visibility: hidden;
+
+		&:checked {
+			+ .label:before {
+				background-color: var(--color-ruby-dark);
+				box-shadow: inset 0 0 0 2px var(--color-secondary-bg);
+			}
+		}
+	}
+
+	.label {
+		display: flex;
+		align-items: center;
+		gap: 7px;
+		transition: .1s ease;
+
+		&:before {
+			content: '';
+			display: block;
+			flex-shrink: 0;
+			width: 15px;
+			height: 15px;
+			border-radius: 50%;
+			border: 1px solid var(--color-black);
+			box-shadow: inset 0 0 0 10px var(--color-secondary-bg);
+			transition: inherit;
+		}
+	}
+}
